+++ Repositories/ComputerRepository.cs
@@ -7,8 +7,9 @@ namespace LabManager.Repositories;
 class ComputerRepository
 {
     private readonly DatabaseConfig _databaseConfig;
-    
-    public ComputerRepository(DatabaseConfig databaseConfig) {
+
+    public ComputerRepository(DatabaseConfig databaseConfig)
+    {
         _databaseConfig = databaseConfig;
     }
 
@@ -47,5 +48,20 @@ class ComputerRepository
         connection.Close();
         
         return computers;
+    }
+
+    public void Save(Computer computer)
+    {
+        var connection = new SqliteConnection(_databaseConfig.ConnectionString);
+        connection.Open();
+
+        var command = connection.CreateCommand();
+        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor)";
+        command.Parameters.AddWithValue("$id", computer.Id);
+        command.Parameters.AddWithValue("$ram", computer.Ram);
+        command.Parameters.AddWithValue("$processor", computer.Processor);
+
+        command.ExecuteNonQuery();
+        connection.Close();
     }
 }
